move ft losses
