+++ sub-packages/bionemo-esm2/src/bionemo/esm2/model/finetune/finetune_token_classifier.py
@@ -15,17 +15,16 @@
 
 
 from dataclasses import dataclass, field
-from typing import Dict, List, Sequence, Tuple, Type
+from typing import List, Sequence, Type
 
 import torch
-from megatron.core import parallel_state
 from megatron.core.transformer.module import MegatronModule
 from megatron.core.transformer.transformer_config import TransformerConfig
 from torch import Tensor
 
 from bionemo.esm2.api import ESM2GenericConfig, ESM2Model
+from bionemo.esm2.model.finetune.loss import ClassifierLossReduction
 from bionemo.llm.model.biobert.model import BioBertOutput
-from bionemo.llm.model.loss import BERTMLMLossWithReduction, PerTokenLossDict, SameSizeLossDict
 from bionemo.llm.utils import iomixin_utils as iom
 
 
@@ -34,68 +33,10 @@ token to output secondary structure predictions.
 """
 
 __all__: Sequence[str] = (
-    "ClassifierLossReduction",
     "MegatronConvNetHead",
     "ESM2FineTuneTokenModel",
     "ESM2FineTuneTokenConfig",
 )
-
-
-class ClassifierLossReduction(BERTMLMLossWithReduction):
-    """A class for calculating the cross entropy loss of classification output.
-
-    This class used for calculating the loss, and for logging the reduced loss across micro batches.
-    """
-
-    def forward(
-        self, batch: Dict[str, Tensor], forward_out: Dict[str, Tensor]
-    ) -> Tuple[Tensor, PerTokenLossDict | SameSizeLossDict]:
-        """Calculates the loss within a micro-batch. A micro-batch is a batch of data on a single GPU.
-
-        Args:
-            batch: A batch of data that gets passed to the original forward inside LitAutoEncoder.
-            forward_out: the output of the forward method inside classification head.
-
-        Returns:
-            A tuple where the loss tensor will be used for backpropagation and the dict will be passed to
-            the reduce method, which currently only works for logging.
-        """
-        targets = batch["labels"].squeeze()  # [b] or [b, s] for sequence-level or token-level classification
-
-        classification_output = forward_out["classification_output"]  # [b, num_class] or [b, s, num_class]
-        # [b, s, num_class] -> [b, num_class, s] to satisfy toke-level input dims for cross_entropy loss
-        if classification_output.dim() == 3:
-            classification_output = classification_output.permute(0, 2, 1)
-
-        loss_mask = batch["loss_mask"]  # [b, s]
-
-        cp_size = parallel_state.get_context_parallel_world_size()
-        if cp_size == 1:
-            losses = torch.nn.functional.cross_entropy(classification_output, targets, reduction="none")
-            # token-level losses may contain NaNs at masked locations. We use masked_select to filter out these NaNs
-            if classification_output.dim() == 3:
-                masked_loss = torch.masked_select(losses, loss_mask)
-                loss = masked_loss.sum() / loss_mask.sum()
-            else:
-                loss = losses.mean()  # sequence-level single value classification
-        else:  # TODO: support CP with masked_token_loss_context_parallel
-            raise NotImplementedError("Context Parallel support is not implemented for this loss")
-
-        return loss, {"avg": loss}
-
-    def reduce(self, losses_reduced_per_micro_batch: Sequence[SameSizeLossDict]) -> Tensor:
-        """Works across micro-batches. (data on single gpu).
-
-        Note: This currently only works for logging and this loss will not be used for backpropagation.
-
-        Args:
-            losses_reduced_per_micro_batch: a list of the outputs of forward
-
-        Returns:
-            A tensor that is the mean of the losses. (used for logging).
-        """
-        losses = torch.stack([loss["avg"] for loss in losses_reduced_per_micro_batch])
-        return losses.mean()
 
 
 class MegatronConvNetHead(MegatronModule):
